// src/enums/Permission.ts
export enum Permission {
	exit = 'exit',
	administrator = 'ADMINISTRATOR',
	createInstantInvite = 'CREATE_INSTANT_INVITE',
	kickMembers = 'KICK_MEMBERS',
	banMembers = 'BAN_MEMBERS',
	manageChannels = 'MANAGE_CHANNELS',
	manageGuild = 'MANAGE_GUILD',
	addReactions = 'ADD_REACTIONS',
	viewAuditLog = 'VIEW_AUDIT_LOG',
	prioritySpeaker = 'PRIORITY_SPEAKER',
	stream = 'STREAM',
	viewChannel = 'VIEW_CHANNEL',
	sendMessages = 'SEND_MESSAGES',
	sendTTSMessages = 'SEND_TTS_MESSAGES',
	manageMessages = 'MANAGE_MESSAGES',
	embedLinks = 'EMBED_LINKS',
	attachFiles = 'ATTACH_FILES',
	readMessageHistory = 'READ_MESSAGE_HISTORY',
	mentionEveryone = 'MENTION_EVERYONE',
	useExternalEmojis = 'USE_EXTERNAL_EMOJIS',
	viewGuildInsights = 'VIEW_GUILD_INSIGHTS',
	connect = 'CONNECT',
	speak = 'SPEAK',
	muteMembers = 'MUTE_MEMBERS',
	deafenMembers = 'DEAFEN_MEMBERS',
	moveMembers = 'MOVE_MEMBERS',
	useVAD = 'USE_VAD',
	changeNickname = 'CHANGE_NICKNAME',
	manageNicknames = 'MANAGE_NICKNAMES',
	manageRoles = 'MANAGE_ROLES',
	manageWebhooks = 'MANAGE_WEBHOOKS',
	manageEmojis = 'MANAGE_EMOJIS'
}